+++ test/integration/src/components/contextual/UserProfile/UserProfile.js
@@ -22,7 +22,7 @@ export default class UserProfile extends DataComponent {
 
     this.subscribe(
       UserSegment.id(), { username: nextProps.username }, 'user',
-      null, hydrationOption);
+      hydrationOption);
   }
 
   render() {